+++ src/hueimmersive/Control.java
@@ -7,42 +7,58 @@ import java.util.TimerTask;
 
 public class Control
 {
+	public static boolean immersiveProcessIsActive = false;
 	private Timer captureLoop;
-	private double autoOffBri = 0.0;
+	private int transitionTime = 5;
+	
+	private float lastAutoOffBri = 0f;
 	
 	public Control() throws Exception
 	{
 		HBridge.setup();
+		
+		if (Main.arguments.contains("force-on") && !Main.arguments.contains("force-off"))
+		{
+			turnAllLightsOn();
+		}
+		if (Main.arguments.contains("force-off") && !Main.arguments.contains("force-on"))
+		{
+			turnAllLightsOff();
+		}
+		if (Main.arguments.contains("force-start"))
+		{
+			startImmersiveProcess();
+		}
 	}
 	
-	public void setLight(int LightID, Color color) throws Exception // calculate color and send it to light
+	public void setLight(HLight light, Color color) throws Exception // calculate color and send it to light
 	{		
-		float[] colorHSB1 = Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null); // unmodified color
+		float[] colorHSB = Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null); // unmodified HSB color
 		
-		color = Color.getHSBColor(colorHSB1[0], colorHSB1[1], (float)(colorHSB1[2] * (Main.ui.slider_Brightness.getValue() / 100.0) * (Settings.Light.getBrightness(LightID) / 100.0))); // calculate brightness set by user
-		float[] colorHSB2 = Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null); // modified color
+		color = Color.getHSBColor(colorHSB[0], Math.max(0f, Math.min(1f, colorHSB[1] * (Main.ui.slider_Saturation.getValue() / 100f))), (float)(colorHSB[2] * (Main.ui.slider_Brightness.getValue() / 100f) * (Settings.Light.getBrightness(light) / 100f))); // modified color
 		
-		double[] xy = HColor.translate(color, Settings.getBoolean("gammacorrection")); // converted color
+		double[] xy = HColor.translate(color, Settings.getBoolean("gammacorrection")); // xy color
+		int bri = Math.round(Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null)[2] * 255); // brightness
 		
-		String APIurl = "http://" + HBridge.internalipaddress + "/api/" + HBridge.username + "/lights/" + LightID + "/state";	
-		String data = "{\"xy\":[" + xy[0] + ", " + xy[1] + "], \"bri\":" + Math.round(colorHSB2[2] * 255) + ", \"transitiontime\":4}"; //\"sat\":" + Math.round(HSBcolor[1] * 255) + ", 
+		String APIurl = "http://" + HBridge.internalipaddress + "/api/" + HBridge.username + "/lights/" + light.id + "/state";	
+		String data = "{\"xy\":[" + xy[0] + ", " + xy[1] + "], \"bri\":" + bri + ", \"transitiontime\":" + transitionTime + "}";
 		
 		// turn light off automatically if the brightness is very low
 		if (Settings.getBoolean("autoswitch"))
 		{
-			if (colorHSB1[2] > autoOffBri + 0.1 && HBridge.getLight(LightID).isOn() == false)
+			if (colorHSB[2] > lastAutoOffBri + 0.1f && light.isOn() == false)
 			{
-				data = "{\"on\":true, \"xy\":[" + xy[0] + ", " + xy[1] + "], \"bri\":" + Math.round(colorHSB2[2] * 255) + ", \"transitiontime\":4}";
+				data = "{\"on\":true, \"xy\":[" + xy[0] + ", " + xy[1] + "], \"bri\":" + bri + ", \"transitiontime\":" + transitionTime + "}";
 			}
-			else if (colorHSB1[2] <= 0.0627451f && HBridge.getLight(LightID).isOn() == true)
+			else if (colorHSB[2] <= 0.0627451f && light.isOn() == true)
 			{
 				data = "{\"on\":false, \"transitiontime\":3}";
-				autoOffBri = colorHSB1[2];
+				lastAutoOffBri = colorHSB[2];
 			}
 		}
-		else if (Settings.getBoolean("autoswitch") == false && HBridge.getLight(LightID).isOn() == false)
+		else if (Settings.getBoolean("autoswitch") == false && light.isOn() == false)
 		{
-			data = "{\"on\":true, \"xy\":[" + xy[0] + ", " + xy[1] + "], \"bri\":" + Math.round(colorHSB2[2] * 255) + ", \"transitiontime\":2}";
+			data = "{\"on\":true, \"xy\":[" + xy[0] + ", " + xy[1] + "], \"bri\":" + bri + ", \"transitiontime\":" + transitionTime + "}";
 		}
 			
 		HRequest.PUT(APIurl, data);
@@ -52,6 +68,12 @@ public class Control
 	{
 		Main.ui.button_Off.setEnabled(false);
 		Main.ui.button_On.setEnabled(false);
+		
+		Main.ui.button_Stop.setEnabled(true);
+		Main.ui.button_Start.setEnabled(false);
+		Main.ui.button_Once.setEnabled(false);
+		
+		immersiveProcessIsActive = true;
 		
 		for(HLight light : HBridge.lights)
 		{
@@ -74,7 +96,7 @@ public class Control
 				}
 			}
 		};
-		captureLoop.scheduleAtFixedRate(task, 0, 300);
+		captureLoop.scheduleAtFixedRate(task, 0, Math.round(transitionTime * 100 * 0.68));
 	}
 	
 	public void stopImmersiveProcess() throws Exception
@@ -82,7 +104,13 @@ public class Control
 		captureLoop.cancel();
 		captureLoop.purge();
 		
+		immersiveProcessIsActive = false;
+		
 		Main.ui.setupOnOffButton();
+		
+		Main.ui.button_Stop.setEnabled(false);
+		Main.ui.button_Start.setEnabled(true);
+		Main.ui.button_Once.setEnabled(true);
 		
 		Thread.sleep(250);
 		ImmersiveProcess.setStandbyOutput();
@@ -99,6 +127,10 @@ public class Control
 	
 	public void onceImmersiveProcess() throws Exception
 	{
+		Main.ui.button_Stop.setEnabled(false);
+		Main.ui.button_Start.setEnabled(true);
+		Main.ui.button_Once.setEnabled(true);
+		
 		ImmersiveProcess.execute();
 	}
 	
@@ -106,22 +138,24 @@ public class Control
 	{
 		for(HLight light : HBridge.lights)
 		{
-			if (Settings.Light.getActive(light.id))
+			if (Settings.Light.getActive(light))
 			{
 				light.turnOn();
 			}
 		}
+		Main.ui.setupOnOffButton();
 	}
 
 	public void turnAllLightsOff() throws Exception
 	{
 		for(HLight light : HBridge.lights)
 		{
-			if (Settings.Light.getActive(light.id))
+			if (Settings.Light.getActive(light))
 			{
 				light.turnOff();
 			}
 		}
+		Main.ui.setupOnOffButton();
 	}
 
 }