+++ src/hueimmersive/HBridge.java
@@ -44,7 +44,7 @@ public class HBridge
 	{
 		Debug.info(null, "try fast connect...");
 		
-		JsonObject response = HRequest.GET("http://" + internalipaddress + "/api/" + username + "/config/");
+		JsonObject response = HRequest.GET("http://" + internalipaddress + "/api/" + username);
 		
 		if (HRequest.responseCheck(response) == "data")
 		{
@@ -57,6 +57,8 @@ public class HBridge
 		}
 		else
 		{
+			Debug.info(null, "can't find bridge");
+			
 			newConnect();
 		}
 	}
@@ -86,6 +88,8 @@ public class HBridge
 						internalipaddress = response.get("internalipaddress").getAsString();
 						
 						Settings.Bridge.setInternalipaddress(internalipaddress);
+						
+						Debug.info(null, "bridge found");
 						
 						login();
 					}